Desktop/Journey/src/requirements.txt
dj-database-url==0.5.0
Django==1.11.17
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-tinymce==2.8.0
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
